added authentication and filled out endpoints

+++ backend/controllers/goalController.js
@@ -1,10 +1,11 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 // GET: /api/goals
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find()
+  const goals = await Goal.find({ user: req.user.id })
   res.json(goals);
 });
 
@@ -16,7 +17,8 @@ const setGoal = asyncHandler(async (req, res) => {
   }
 
   const goal = await Goal.create({
-    text: req.body.text
+    text: req.body.text,
+    user: req.user.id
   })
 
   res.json(goal);
@@ -32,6 +34,19 @@ const updateGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
 
+  const user = await User.findById(req.user.id);
+
+  // Next two code blocks seem useless, don't need to check user and don't need to query for user to use id again
+  if(!user){
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(goalId, { text: req.body.text }, { new: true })
 
   res.json(updatedGoal);
@@ -45,6 +60,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if(!goal){
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // Next two code blocks seem useless, don't need to check user and don't need to query for user to use id again
+  if(!user){
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await Goal.findByIdAndRemove(goalId)
